fr_cli/requirements.txt
zhipuai>=2.0.0
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.0
pymysql>=1.0.0
psycopg2-binary>=2.9.0
pyodbc>=4.0.0
oracledb>=1.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
paramiko>=3.0.0
selenium>=4.10.0
watchdog>=3.0.0

# 以下为可选云存储依赖，按需安装：
# bypy (百度网盘)
# aligo (阿里云盘)
# msal (OneDrive)
